Issue-#29: Refactor test structure.
